# server/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: felixtj.dietrich@tum.de
    name: Felix T.J. Dietrich
  description: API documentation for the Hephaestus application server.
  license:
    name: MIT License
    url: https://github.com/ls1intum/Hephaestus/blob/develop/LICENSE
  title: Hephaestus API
  version: 0.0.0-development
servers:
- description: Default Server URL
  url: /
security:
- bearerAuth: []
...
paths:
...
  /workspaces/{workspaceSlug}/team/{id}/visibility:
    post:
      description: Show or hide a team in leaderboard calculations
      operationId: updateTeamVisibility
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      - description: "whether to hide the team (query parameter, fallback)"
        in: query
        name: hidden
        required: false
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: boolean
        description: "whether to hide the team (body parameter, preferred)"
      responses:
        "200":
          description: "200 OK on success, 404 if team not found"
      summary: Update team visibility
      tags:
      - Teams
  /workspaces/{workspaceSlug}/team/{teamId}/repositories/{repositoryId}/visibility:
    post:
      operationId: updateRepositoryVisibility
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: hiddenFromContributions
        required: false
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: boolean
      responses:
        "200":
          description: OK
      tags:
      - Teams
...
  /workspaces/{workspaceSlug}/teams/{teamId}/settings:
    get:
      description: Returns the visibility settings for a team in the workspace
      operationId: getTeamSettings
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceTeamSettings"
          description: Team settings returned
      security: []
      summary: Get team settings
      tags:
      - Team Settings
    patch:
      description: Updates the visibility settings for a team in the workspace
      operationId: updateTeamSettings
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamSettingsRequest"
        description: the update request containing the hidden flag
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceTeamSettings"
          description: Team settings updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceTeamSettings"
          description: Team not found in workspace
      summary: Update team settings
      tags:
      - Team Settings
  /workspaces/{workspaceSlug}/teams/{teamId}/settings/label-filters:
    get:
      description: Returns all labels configured as filters for a team in the workspace
      operationId: getLabelFilters
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabelInfo"
          description: Label filters returned
      security: []
      summary: Get team label filters
      tags:
      - Team Settings
  /workspaces/{workspaceSlug}/teams/{teamId}/settings/label-filters/{labelId}:
    delete:
      description: Removes a label filter from a team in the workspace
      operationId: removeLabelFilter
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - description: the label ID to remove as filter
        in: path
        name: labelId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Label filter removed
        "404":
          description: Label filter not found
      summary: Remove label filter
      tags:
      - Team Settings
    post:
      description: Adds a label as a filter for a team in the workspace
      operationId: addLabelFilter
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - description: the label ID to add as filter
        in: path
        name: labelId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Label filter added
        "404":
          description: Team or label not found
      summary: Add label filter
      tags:
      - Team Settings
  /workspaces/{workspaceSlug}/teams/{teamId}/settings/repositories/{repositoryId}:
    get:
      description: Returns the contribution visibility settings for a repository in
        a team
      operationId: getRepositorySettings
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - description: the repository ID
        in: path
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceTeamRepositorySettings"
          description: Repository settings returned
      security: []
      summary: Get repository settings
      tags:
      - Team Settings
    patch:
      description: Updates the contribution visibility settings for a repository in
        a team
      operationId: updateRepositorySettings
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - description: the team ID
        in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - description: the repository ID
        in: path
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRepositorySettingsRequest"
        description: the update request containing the hiddenFromContributions flag
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceTeamRepositorySettings"
          description: Repository settings updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceTeamRepositorySettings"
          description: Team or repository not found
      summary: Update repository settings
      tags:
      - Team Settings
...
components:
  schemas:
...
    LabelInfo:
      type: object
      description: Information about a label from a repository
      properties:
        color:
          type: string
          description: "Hex color code of the label (without #)"
          example: d73a4a
        id:
          type: integer
          format: int64
          description: Unique identifier of the label
        name:
          type: string
          description: Name of the label
        repository:
          $ref: "#/components/schemas/RepositoryInfo"
          description: Repository the label belongs to
      required:
      - color
      - id
      - name
...
    RepositoryInfo:
      type: object
      description: Information about a git repository
      properties:
        description:
          type: string
          description: Description of the repository
        hiddenFromContributions:
          type: boolean
          description: Whether contributions from this repository are hidden from
            leaderboard calculations
        htmlUrl:
          type: string
          description: URL to the repository on the git provider
        id:
          type: integer
          format: int64
          description: Unique identifier of the repository
        labels:
          type: array
          description: Labels defined in the repository
          items:
            $ref: "#/components/schemas/LabelInfo"
        name:
          type: string
          description: Name of the repository
        nameWithOwner:
          type: string
          description: "Full name including owner (e.g., 'owner/repo')"
      required:
      - hiddenFromContributions
      - htmlUrl
      - id
      - name
      - nameWithOwner
...
    UpdateRepositorySettingsRequest:
      type: object
      description: Request to update repository contribution visibility settings in
        a workspace
      properties:
        hiddenFromContributions:
          type: boolean
          description: Whether contributions from this repository should be hidden
            from leaderboard calculations
      required:
      - hiddenFromContributions
...
    UpdateTeamSettingsRequest:
      type: object
      description: Request to update team visibility settings in a workspace
      properties:
        hidden:
          type: boolean
          description: Whether the team should be hidden from the leaderboard
      required:
      - hidden
...
    WorkspaceTeamRepositorySettings:
      type: object
      description: Repository contribution visibility settings for a specific team
        within a workspace
      properties:
        hiddenFromContributions:
          type: boolean
          description: Whether contributions from this repository are hidden from
            leaderboard calculations
        repositoryId:
          type: integer
          format: int64
          description: The repository ID these settings apply to
        teamId:
          type: integer
          format: int64
          description: The team ID these settings apply to
        workspaceId:
          type: integer
          format: int64
          description: The workspace ID these settings belong to
      required:
      - hiddenFromContributions
      - repositoryId
      - teamId
      - workspaceId
    WorkspaceTeamSettings:
      type: object
      description: Team visibility settings for a specific workspace
      properties:
        hidden:
          type: boolean
          description: Whether the team is hidden in the leaderboard for this workspace
        teamId:
          type: integer
          format: int64
          description: The team ID these settings apply to
        workspaceId:
          type: integer
          format: int64
          description: The workspace ID these settings belong to
      required:
      - hidden
      - teamId
      - workspaceId
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Hephaestus-native JWT bearer authentication. The SPA normally authenticates
        via the `__Host-HEPHAESTUS_AT` session cookie; this scheme documents the equivalent
        bearer token.
      scheme: bearer
      type: http